Option to validate authenticated user if part of an organization
Users in GitHub can be part of organizations, and so we provide an option to
validate if a user is a member of an organization before successfully
authenticating.  Organization is set through the securesocial.github.organization
key in application.conf.

Note that a user's membership should be publicized for it to be identified
as part of a given organization.

Signed-off-by: Ealden Esto E. Escanan <ealden@gmail.com>