Make SecureSocial mostly compile with Play 2.1-RC1
Break out code that waits on HTTP requests for auth token and user details
to IdentityProvider.awaitResultOrThrowAuthnException().

This commit is actually a fix for an *older* version of SecureSocial (namely
commit 7ec359cc97c4 from 2012-06-22) that I'm now rebasing on SecureSocial's
current master head. Since this is a rebase of a somewhat old commit, it won't
make SS work completely with Play 2.1 and there'll be compilation errors; I'll
fix them in the next few commits.