use query parameter to pass referer
The HTTP header "referer" is not always set. In our
case, Heroku did not pass the header to the 
application.

Solution is now that login first checks whether
a query parameter with the referer is passed, if
that is not the case the HTTP header "referer" is
checked, and only if both fail the usual fallback
is triggered.