Add onPasswordChangeGoTo option
Adds the new, optional 'securesocial.onPasswordChangeGoTo'
configuration property to allow the developer to set which page
the Password Change form should redirect to.

If not set, it preserves the current default option of staying in the
Password Change page.

This option is similar to existing options such as onSignUpGoTo and
onResetPasswordGoTo.