Migrate to using named loggers
Using named loggers allows client applications to have a fine grained control over the library logs. An example could be logging all sent emails for tracking purpose without setting all the application's logs at DEBUG which can be extremely verbose.

This commit names all the loggers used throughout securesocial, adds loggers to the sample app to demonstrate that there are no conflicts in defining a logger variable with the same name in the application code. Finally it changes the sample applications configuration to show how to set the log level for the securesocial loggers.