Replace hardcoded async execution contexts with a configurable parameter
The default execution context (scala.concurrent.ExecutionContext.global)
was hardcoded in multiple places in the code. This is not only
inflexible, but problematic for Play applications, since it will cause
code inside futures to not use the expected environment, like the custom
class loader.

As as solution, introduce an executionContext member to
RuntimeEnvironment, defaulting to the global context, that can be
overridden by users as desired.

Services that need the contexts have been refactored to receive them
once as a constructor parameter in their default implementations, or
whenever possible, take them from another parameter that already has one
(like another service). The only service that was not linked to any
other and needs it's own definition is AuthenticatorStore.
To mitigate issues, it will get a context from the
environment in the default implementation (through an implicit
parameter), or default to the global context if no other override
exists.