enable use of referer if user is already authenticated
The following use-case was not supported so far and can happen in a
productive environment with enabled HTTPS support.

1. User authenticates
2. User goes to a resource using http (instead of https) by either
entering a URL or by clicking on an absolute link. Let’s say
http://domain.com/some/public/resource
3. Since HTTP is used, the user is not authenticated. User clicks on
“Login” to authenticate
4. Since the user is already authenticated from step 1. he/she is
redirected to /. Expected would be that the user is redirected to
https://domain.com/some/public/resource

Fix:
There already was a feature that the referer is used (see config
securesocial.SecureSocial.enableRefererAsOriginalUrl). This feature was
only applied if the user was not yet authenticated. This commit enables
the same logic for authenticated users.

Testing:
Make sure the feature is enabled (config property `securesocial.
enableRefererAsOriginalUrl = true`)
1.  go to http://localhost/current-link
2. authenticate
3. click on “login” to do a second login attempt
4. you should be redirected to /current-link again